pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "groq>=0.28.0",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]
